4-Gitflow: - Added Spring Boot and Lombok
