+++ src/main/java/ru/internship/mvc/MainAppStart.java
@@ -1,0 +1,12 @@
+package ru.internship.mvc;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MainAppStart {
+
+    public static void main(String[] args) {
+        SpringApplication.run(MainAppStart.class, args);
+    }
+}
